処理の修正
